Before batch change using copilot
