+++ app/Models/Context.php
@@ -39,7 +39,7 @@ class Context extends Model
     ];
 
     // Add a method to mark a single row as the default
-    public function markAsDefault()
+    public function setDefault()
     {
         // Ensure the table has a 'default' column (boolean or integer)
         if (! $this->exists) {
