+++ tests/Feature/Health/Test.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Health;
 
 use App\Models\User;
 use Tests\TestCase;
 
-class HealthTest extends TestCase
+class Test extends TestCase
 {
     public function test_the_application_is_up_and_running_as_anonymous(): void
     {
